# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MOTION Workshop Manager",
    "description": "Backend API for MOTION SaaS",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/jobs/": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create Job Card",
        "operationId": "create_job_card_api_v1_jobs__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCardCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_api_v1_jobs__get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobCardResponse"
                  },
                  "title": "Response List Jobs Api V1 Jobs  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "jobs"
        ],
        "summary": "Update Job",
        "operationId": "update_job_api_v1_jobs__job_id__patch",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCardUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/parts": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Add Job Part",
        "operationId": "add_job_part_api_v1_jobs__job_id__parts_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobPartCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/labor": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Add Job Labor",
        "operationId": "add_job_labor_api_v1_jobs__job_id__labor_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobLaborCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobLaborResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/invoices/": {
      "post": {
        "tags": [
          "invoices"
        ],
        "summary": "Create Invoice",
        "operationId": "create_invoice_api_v1_invoices__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "invoices"
        ],
        "summary": "List Invoices",
        "operationId": "list_invoices_api_v1_invoices__get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceResponse"
                  },
                  "title": "Response List Invoices Api V1 Invoices  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoices/{invoice_id}": {
      "get": {
        "tags": [
          "invoices"
        ],
        "summary": "Get Invoice",
        "operationId": "get_invoice_api_v1_invoices__invoice_id__get",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoices/{invoice_id}/items": {
      "post": {
        "tags": [
          "invoices"
        ],
        "summary": "Add Invoice Item",
        "operationId": "add_invoice_item_api_v1_invoices__invoice_id__items_post",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceLineItemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceLineItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvoiceCreate": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "invoice_number": {
            "type": "string",
            "maxLength": 50,
            "title": "Invoice Number"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "status": {
            "type": "string",
            "maxLength": 50,
            "title": "Status",
            "default": "DRAFT"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "terms": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "invoice_number",
          "tenant_id"
        ],
        "title": "InvoiceCreate"
      },
      "InvoiceLineItemCreate": {
        "properties": {
          "job_part_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Part Id"
          },
          "job_labor_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Labor Id"
          },
          "description": {
            "type": "string",
            "maxLength": 255,
            "title": "Description"
          },
          "quantity": {
            "type": "number",
            "minimum": 0.0,
            "title": "Quantity",
            "default": 1.0
          },
          "unit_price": {
            "type": "number",
            "minimum": 0.0,
            "title": "Unit Price",
            "default": 0.0
          },
          "tax_rate": {
            "type": "number",
            "minimum": 0.0,
            "title": "Tax Rate",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "description"
        ],
        "title": "InvoiceLineItemCreate"
      },
...
      "InvoiceResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "invoice_number": {
            "type": "string",
            "maxLength": 50,
            "title": "Invoice Number"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "status": {
            "type": "string",
            "maxLength": 50,
            "title": "Status",
            "default": "DRAFT"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "terms": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "issue_date": {
            "type": "string",
            "format": "date-time",
            "title": "Issue Date"
          },
          "subtotal": {
            "type": "number",
            "title": "Subtotal"
          },
          "total_tax": {
            "type": "number",
            "title": "Total Tax"
          },
          "total_amount": {
            "type": "number",
            "title": "Total Amount"
          },
          "amount_paid": {
            "type": "number",
            "title": "Amount Paid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "line_items": {
            "items": {
              "$ref": "#/components/schemas/InvoiceLineItemResponse"
            },
            "type": "array",
            "title": "Line Items",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "invoice_number",
          "id",
          "tenant_id",
          "issue_date",
          "subtotal",
          "total_tax",
          "total_amount",
          "amount_paid",
          "created_at",
          "updated_at"
        ],
        "title": "InvoiceResponse"
      },
      "JobCardCreate": {
        "properties": {
          "booking_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Booking Id"
          },
          "vehicle_make": {
            "type": "string",
            "maxLength": 100,
            "title": "Vehicle Make"
          },
          "vehicle_model": {
            "type": "string",
            "maxLength": 100,
            "title": "Vehicle Model"
          },
          "registration_number": {
            "type": "string",
            "maxLength": 50,
            "title": "Registration Number"
          },
          "mileage": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Mileage"
          },
          "status": {
            "type": "string",
            "maxLength": 50,
            "title": "Status",
            "default": "DRAFT"
          },
          "customer_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Notes"
          },
          "mechanic_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mechanic Notes"
          },
          "assigned_mechanic_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Mechanic Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          }
        },
        "type": "object",
        "required": [
          "vehicle_make",
          "vehicle_model",
          "registration_number",
          "tenant_id"
        ],
        "title": "JobCardCreate"
      },
      "JobCardResponse": {
        "properties": {
          "booking_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Booking Id"
          },
          "vehicle_make": {
            "type": "string",
            "maxLength": 100,
            "title": "Vehicle Make"
          },
          "vehicle_model": {
            "type": "string",
            "maxLength": 100,
            "title": "Vehicle Model"
          },
          "registration_number": {
            "type": "string",
            "maxLength": 50,
            "title": "Registration Number"
          },
          "mileage": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Mileage"
          },
          "status": {
            "type": "string",
            "maxLength": 50,
            "title": "Status",
            "default": "DRAFT"
          },
          "customer_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Notes"
          },
          "mechanic_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mechanic Notes"
          },
          "assigned_mechanic_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Mechanic Id"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "parts": {
            "items": {
              "$ref": "#/components/schemas/JobPartResponse"
            },
            "type": "array",
            "title": "Parts",
            "default": []
          },
          "labor": {
            "items": {
              "$ref": "#/components/schemas/JobLaborResponse"
            },
            "type": "array",
            "title": "Labor",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "vehicle_make",
          "vehicle_model",
          "registration_number",
          "id",
          "tenant_id",
          "created_at",
          "updated_at"
        ],
        "title": "JobCardResponse"
      },
      "JobCardUpdate": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "mileage": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Mileage"
          },
          "customer_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Notes"
          },
          "mechanic_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mechanic Notes"
          },
          "assigned_mechanic_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Mechanic Id"
          }
        },
        "type": "object",
        "title": "JobCardUpdate"
      },
      "JobLaborCreate": {
        "properties": {
          "description": {
            "type": "string",
            "maxLength": 255,
            "title": "Description"
          },
          "hours": {
            "type": "number",
            "minimum": 0.0,
            "title": "Hours",
            "default": 1.0
          },
          "hourly_rate": {
            "type": "number",
            "minimum": 0.0,
            "title": "Hourly Rate",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "description"
        ],
        "title": "JobLaborCreate"
      },
      "JobLaborResponse": {
        "properties": {
          "description": {
            "type": "string",
            "maxLength": 255,
            "title": "Description"
          },
          "hours": {
            "type": "number",
            "minimum": 0.0,
            "title": "Hours",
            "default": 1.0
          },
          "hourly_rate": {
            "type": "number",
            "minimum": 0.0,
            "title": "Hourly Rate",
            "default": 0.0
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "description",
          "id",
          "job_id"
        ],
        "title": "JobLaborResponse"
      },
      "JobPartCreate": {
        "properties": {
          "part_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Number"
          },
          "description": {
            "type": "string",
            "maxLength": 255,
            "title": "Description"
          },
          "quantity": {
            "type": "number",
            "minimum": 0.0,
            "title": "Quantity",
            "default": 1.0
          },
          "unit_cost": {
            "type": "number",
            "minimum": 0.0,
            "title": "Unit Cost",
            "default": 0.0
          },
          "unit_price": {
            "type": "number",
            "minimum": 0.0,
            "title": "Unit Price",
            "default": 0.0
          },
          "is_supplied_by_customer": {
            "type": "boolean",
            "title": "Is Supplied By Customer",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "description"
        ],
        "title": "JobPartCreate"
      },
      "JobPartResponse": {
        "properties": {
          "part_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Number"
          },
          "description": {
            "type": "string",
            "maxLength": 255,
            "title": "Description"
          },
          "quantity": {
            "type": "number",
            "minimum": 0.0,
            "title": "Quantity",
            "default": 1.0
          },
          "unit_cost": {
            "type": "number",
            "minimum": 0.0,
            "title": "Unit Cost",
            "default": 0.0
          },
          "unit_price": {
            "type": "number",
            "minimum": 0.0,
            "title": "Unit Price",
            "default": 0.0
          },
          "is_supplied_by_customer": {
            "type": "boolean",
            "title": "Is Supplied By Customer",
            "default": false
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "description",
          "id",
          "job_id"
        ],
        "title": "JobPartResponse"
      },
...
    }
  }
}